requirements.txt
numpy
pandas
scikit-learn
mlflow
pyyaml
from_root
boto3
mysql
pypiwin32